Move label input to its own component
