+++ core/src/app/app.module.ts
@@ -118,6 +118,7 @@ import { LimitRangeHeaderRendererComponent } from './content/environments/config
 import { Copy2ClipboardModalComponent } from './shared/components/copy2clipboard-modal/copy2clipboard-modal.component';
 import { LoginErrorComponent } from './content/login-error/login-error.component';
 import { CreateRemoteEnvironmentModalComponent } from './content/settings/remote-environments/create-remote-environment-modal/create-remote-environment-modal.component';
+import { LabelsInputComponent } from './shared/components/labels-input/labels-input.component';
 
 @NgModule({
   declarations: [
@@ -204,7 +205,8 @@ import { CreateRemoteEnvironmentModalComponent } from './content/settings/remote
     LimitRangeEntryRendererComponent,
     LimitRangeHeaderRendererComponent,
     LoginErrorComponent,
-    CreateRemoteEnvironmentModalComponent
+    CreateRemoteEnvironmentModalComponent,
+    LabelsInputComponent
   ],
   imports: [
     BrowserModule,
