+++ core/src/app/content/settings/remote-environments/create-remote-environment-modal/create-remote-environment-modal.component.html
@@ -34,20 +34,7 @@
             </div>
             <div class="tn-form__item">
               <label class="tn-form__label" for="input-2">Labels</label>
-              <div class="labels-input" id="input-2">
-                <span title="Edit label" class="tn-badge-labels tn-badge tn-badge--pill labels" *ngFor="let label of labels"
-                  (click)="updateLabel(label)">
-                  {{ label }}
-                  <div class="tn-icon tn-icon--remove remove-btn labels__remove" (click)="removeLabel(label)"></div>
-                </span>
-                <input #labelsInput type="text" class="tn-form__control" [(ngModel)]="newLabel" (keydown.enter)="addLabel()" (keydown.tab)="addLabel()"
-                  (blur)="addLabel()"
-                  [ngClass]="{'has-labels': labels.length > 0}"
-                  placeholder="{{labels.length === 0 ? 'Specify one or more labels for your new Remote Environment' : ''}}">
-              </div>
-              <div *ngIf="wrongLabelMessage" class="alert alert-danger">
-                {{ wrongLabelMessage }}
-              </div>
+              <app-labels-input (labelsChangeEmitter$)="updateLabelsData($event)"></app-labels-input>
             </div>
           </div>
         </div>
