+++ core/src/app/content/settings/remote-environments/create-remote-environment-modal/create-remote-environment-modal.component.scss
@@ -9,56 +9,3 @@
 .tn-form__control::placeholder {
   font-style: italic;
 }
-
-
-.labels-input {
-  border: 1px solid #d7d7d7;
-  border-radius: 4px;
-  width: 100%;
-  height: auto;
-  padding: 6px 0;
-  overflow-x: hidden;
-  overflow-y: hidden; // white-space: nowrap;
-  &:hover {
-    border: 1px solid #73B9FF;
-  }
-
-  input {
-    border: none;
-    position: relative;
-    height: auto;
-    width: 100%;
-    &.has-labels {
-      width: auto;
-    }
-  }
-
-  .labels {
-    position: relative;
-    bottom: 2px;
-    &:first-child {
-      margin-left: 8px;
-    }
-    margin-right: 5px;
-    padding-right: 25px;
-    &__remove {
-      font-size: .5rem;
-      margin-left: 5px;
-      position: absolute;
-      cursor: pointer;
-      top: 50%;
-      right: 6px;
-      transform: translateY(-50%);
-      transition: 0.2s linear font-size;
-    }
-    &:hover {
-      .labels__remove {
-        font-size: 0.7rem;
-        transition: 0.2s linear font-size;
-      }
-      background-color: darken(#515559, 20%);
-      cursor: pointer;
-    }
-  }
-}
-
